requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
# python-telegram-bot==13.15  # Removed due to compatibility issues
google-generativeai==0.3.1
qrcode==7.4.2
Pillow==10.1.0
stripe==7.5.0
passlib==1.7.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.23
pyjwt==2.8.0
gunicorn==21.2.0